apps/api/pyproject.toml
[project]
name = "tailorflow-api"
version = "0.1.0"
description = "TailorFlow backend API – FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.27,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "sqlalchemy[asyncio]>=2,<3",
    "asyncpg>=0.29,<1",
    "alembic>=1.13,<2",
    "python-jose[cryptography]>=3.3,<4",
    "bcrypt>=4.1,<6",
    "email-validator>=2,<3",
    "python-multipart>=0.0.6",
    "python-dotenv>=1,<2",
    "openai>=1.10,<2",
    "python-docx>=1,<2",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "httpx>=0.27,<1",
    "aiosqlite>=0.20,<1",
    "ruff>=0.3,<1",
    "mypy>=1.8,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../../tests/backend"]

[tool.mypy]
python_version = "3.11"
strict = true
